renamed utility modules to util
