+++ tests/asdf_tests/test_asdf_core.py
@@ -14,7 +14,7 @@ from scipy.spatial.transform import Rotation
 import weldx.transformations as tf
 from tests._helpers import get_test_name
 from weldx.asdf.tags.weldx.core.file import ExternalFile
-from weldx.asdf.utils import _write_buffer, _write_read_buffer
+from weldx.asdf.util import _write_buffer, _write_read_buffer
 from weldx.constants import WELDX_QUANTITY as Q_
 from weldx.core import MathematicalExpression as ME  # nopep8
 from weldx.core import TimeSeries
