+++ weldx/asdf/util.py
