+++ weldx/util.py
